# swagger_doc.yaml
openapi: 3.0.3
servers:
  - url: http://localhost:3000
info:
  title: Module 295 ToDo API System
  description: |
    A service which stores tasks for users, as well as include a login system for them so they can have their individual workspaces. 
  contact:
    name: Manassés Zähnler
    email: manasses.zaehnler@outlook.com
  version: 1.0.0

tags:
  - name: todo
  - name: tasks

paths:
  '/tasks':
    get:
      tags:
        - tasks
      summary: Read all tasks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/task'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403Unauthorized'

    post: 
      tags:
        - tasks
      summary: Create a new task
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/task'
      responses:  
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/task'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403Unauthorized'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Missing name property"
    
  '/tasks/{id}':
    get:
      tags:
        - tasks
      summary: Read a single task by its id
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/task'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403Unauthorized'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404TaskNotFound'

    put:
      tags: 
        - tasks
      summary: Update a task by its id
      parameters: 
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/task'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/task'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403Unauthorized'
        '403 (No permission)':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403CannotChangeTask'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404TaskNotFound'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "Missing or wrong body properties"

    delete:
      tags: 
        - tasks
      summary: Delete a task by its id
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: No Content
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403CannotChangeTask'
        '403 (No permission)':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403CannotChangeTask'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404TaskNotFound'
          
  
components:
  parameters:    
    id:
      name: id
      in: path
      required: true
      description: The id of a task
      schema:
        type: integer

  schemas:
    task:
      description: A task entry
      type: object
      properties:
        id:
          type: integer
          example: 21
        name: 
          type: string
          example: "Do joga"
        created_at:
          type: date
          example: "2023-06-08T00:00:00.000Z"
        completed_at:
          type: date 
          example: "2023-06-10T00:00:00.000Z"
        email:
          type: string
          example: "jane.smith@gmail.com"

    403Unauthorized:
      description: Thrown when unauthorized
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"

    403CannotChangeTask:
      description: Thrown when trying to change a task the user does not have power over
      type: object
      properties:
        error:
          type: string
          example: "Cannot change a task you do not own"

    404TaskNotFound:
      description: Thrown when task not found
      type: object
      properties:
        error:
          type: string
          example: "Task not found"